DBUG weighting (use normal distribution)
The use of a SCGD is not necessary since we only want relative weights
of >0 forecasted precipitation.
The use of a gamma distribution is much more easier and avoids the
problem of the "near-zero" weights (the shift of the SCGD lead to
incoherent weighting for observation values near 0mm, which resulted in
underestimation of precipitation after assimilationi, especially at an
hourly time step).

This commit also add some visualization tools (temporal evalutions,
focus on a specific pixel,...)