Add possibility to run assimilation on a set of point only
The goal is to reduce the running time by working only on points where
nivometeo observations are available for evaluation over the Alps.

The main probleme is the dimension of the various assimilation fields (3
diensions instead of 4 : (lon, lat) --> 'num_poste'

Another major problem was the reading of pearome data over the Alps
domain wich saturated memory. The solution was to keep the data in a
dask.array state until the temporal loop allows to reduce data size
considerably.