Read large hourly ensemble data
When the domain is large (for example the french Alps : ~120*150 grid cells) the computation is complex to avoid memory crashes.

The "open_mfdataset" method uses dask parallelisation to optimise computation time. It breaks up the data into "chunks" so that
operations can be parallelized more efficiently. By default, chunks will be chosen to load entire input files into memory at once,
which results in memory crashes.
It is advised to specify small chunks, but it is not possible to chunk
on lat/lon coordinates since there is an interpolation step early in the
algorithm.

The first working solution is to use "24 time step" chunks but it can
certainly be optimised.