Add new mask and ratio estimation based on local homogeneity of precipitation accumulation
Scores are projected in space according to the proximity of
precipitation accumulation under the assumption that 2 similar accumulation
not too far away have similar biases.
==> This bias estimation is used for both debiasing and increasing the
observation error.depending on the precipitaion value

Similarly a homogeneity analysis is used for a static increase of the observation
error where the heterogeneity of precipitation accumulation is
important.

Also modify figures to improve visualisation and presentations.