New observation error and debiaising fields
Two steps :
1. filter accumulation field to produce a map of deviation to the
smoothed field (parameter : moving window size set to 15)
2. Use evaluation data to apply the ratio to neighboring points with a
ponderation depending on the ratio estimated at step 1 and the distance
between each pixel and the scores (parameter : correlation lenght set to 0.15)

This provides both a debiaising mask to apply to each new observation
before assimilation and an observation error field.

When this debiasing method is applied to ANTILOPE accumulations, the
field is smoother and visible artifical patterns are attenuated.