Update method
The main problem with the previous method is that the sum of the weights
can be >1.
The solution proposed in this commit is to consider only the stations
with the greater weights up to a total weight of 1.
For points were the sum of the weights is still <1, the weight given to
the default ratio (1) is 1-sum(weight).