Use all evaluation data for the ratio estimation
Give a weight to the default ratio (1) depending on the sum of the
weights of the different evaluation stations:
- 0 if total weight >= 1
- 1 - total weight else

The result seem more efficient by computing weights with an exponential
decrease instead of a normal one.

First evaluation indicates that increasing c0 reduces the weight of >0
biases and that increasing d0 reduces the bias correction over the Mt
Blanc.

So far the best configuration seem to be c0=2 and d0=0.2.