Compute a distribution independent for each ensemble member
The goal is to update background values by taking into account
surrounding values.
This allow to account for correlations and still deal with diagonal
ECMs.

The idea is to estimate the background distribution for each pixel of
each memeber considering the weighted values in surrounding pixels.

The final ECM value is computed mixing weighted values of all members
and all surrounding pixels.