DBUG ECM computation
In the previous commit the background ECM was the sum of 16 standard
deviations, dividing by 16 produces Background ACM and Obseration
dynaimic ECM more comparable (still the dynamic error is significantly
lower but the addition of static error makes up for that).

DBUG : Rstat is the square root of sigma (everything must be in the
R^1/2 space)

Clarify the different method implemented in the comments

TODO : At this point the main remaining issue is the smoothing of the various
fields by pulling original values toward the local weighted average.