Major modification of the algorithm
The ratio estimation now includes a measure of the uncertainty of the
method (spread of the set of estiamted ratios with the different available
reference points).
The final ratio estimation is pulled toward the mean value between 1 and
the mean estimated ratio when the spread increases.

This leads to smoother ratio fields with an average bias (over the
nivometeo stations) < 0.1.

This ratio estimation aims at identifying large systematic error areas (such as
mountain ridges), not to introduce biases where errors are more random.